refactor: configuration 패키지명 변경
